requirements.txt
pytesseract
tox
python-Levenshtein